Organize authentication - Move auth service,controller to client module - Move redis service to shared module - Move email service to shared module
