+++ src/app.module.ts
@@ -4,10 +4,8 @@ import { AppService } from "./app/app.service";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { config } from "./common/Config";
 import { ProductEntity } from "./shared/products/entities/Product.entity";
-import { AuthController } from "./auth/auth.controller";
-import { AuthService } from "./auth/auth.service";
 import { UserEntity } from "./users/entities/User.entity";
-import { RedisService } from "./redis/redis.service";
+import { RedisService } from "./shared/redis/redis.service";
 import { UsersController } from "./users/users.controller";
 import { UsersService } from "./users/users.service";
 import { AddressEntity } from "./users/entities/Address.entity";
@@ -21,7 +19,7 @@ import { CardController } from "./payments/card/card.controller";
 import { CardService } from "./payments/card/card.service";
 import { RavepayService } from "./ravepay/ravepay.service";
 import { NotificationsService } from "./notifications/notifications.service";
-import { EmailService } from "./email/email.service";
+import { EmailService } from "./shared/email/email.service";
 import { AdminModule } from "./admin/admin.module";
 import { SharedModule } from "./shared/shared.module";
 import { ClientModule } from "./client/client.module";
@@ -53,7 +51,6 @@ import * as path from "path";
   ],
   controllers: [
     AppController,
-    AuthController,
     UsersController,
     OrdersController,
     MpesaController,
@@ -61,7 +58,6 @@ import * as path from "path";
   ],
   providers: [
     AppService,
-    AuthService,
     RedisService,
     UsersService,
     OrdersService,
