--- src/auth/auth.controller.spec.ts
@@ -1,39 +1,0 @@
-import { Test, TestingModule } from "@nestjs/testing";
-import { AuthController } from "./auth.controller";
-import { AuthService } from "./auth.service";
-import { getRepositoryToken } from "@nestjs/typeorm";
-import { UserEntity } from "../users/entities/User.entity";
-import { Repository } from "typeorm";
-import { RedisService } from "../redis/redis.service";
-import { EmailService } from "../email/email.service";
-
-describe("Auth Controller", () => {
-  let controller: AuthController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [AuthController],
-      providers: [
-        AuthService,
-        {
-          provide: getRepositoryToken(UserEntity),
-          useClass: Repository,
-        },
-        {
-          provide: RedisService,
-          useClass: jest.fn(),
-        },
-        {
-          provide: EmailService,
-          useClass: jest.fn(),
-        },
-      ],
-    }).compile();
-
-    controller = module.get<AuthController>(AuthController);
-  });
-
-  it("should be defined", () => {
-    expect(controller).toBeDefined();
-  });
-});
