+++ src/client/auth/auth.controller.ts
@@ -22,10 +22,10 @@ import { FastifyReply } from "fastify";
 import { ServerResponse } from "http";
 import { LoginResponseDto } from "./dto/LoginResponse.dto";
 import { SignUpDto } from "./dto/SignUp.dto";
-import { ApiResponseDto } from "../common/dto/ApiResponse.dto";
+import { ApiResponseDto } from "../../common/dto/ApiResponse.dto";
 import { SignInDto } from "./dto/SignIn.dto";
 import * as moment from "moment";
-import { AuthGuard } from "../common/guards/auth.guard";
+import { AuthGuard } from "../../common/guards/auth.guard";
 import { ChangePasswordDto } from "./dto/ChangePassword.dto";
 import { RequestResetPasswordDto } from "./dto/RequestResetPassword.dto";
 import { ResetPasswordDto } from "./dto/ResetPassword.dto";
