+++ src/client/auth/dto/ResetPasswordToken.dto.ts
