+++ src/main.ts
@@ -12,7 +12,7 @@ import {
 } from "@nestjs/platform-fastify";
 import * as fastifyRateLimit from "fastify-rate-limit";
 import * as fileUpload from "fastify-file-upload";
-import { RedisService } from "./redis/redis.service";
+import { RedisService } from "./shared/redis/redis.service";
 import { NotificationsService } from "./notifications/notifications.service";
 
 async function bootstrap() {
