+++ src/payments/mpesa/mpesa.service.ts
@@ -2,12 +2,12 @@ import { Injectable, HttpStatus } from "@nestjs/common";
 import { PaymentEntity } from "../entities/Payment.entity";
 import { InjectRepository } from "@nestjs/typeorm";
 import { Repository } from "typeorm";
-import { RedisService } from "../../redis/redis.service";
+import { RedisService } from "../../shared/redis/redis.service";
 import { config } from "../../common/Config";
 import * as axios from "axios";
 import * as moment from "moment";
 import { RequestMpesaDto } from "./dto/RequestMpesa.dto";
-import { SessionDto } from "../../auth/dto/Session.dto";
+import { SessionDto } from "../../client/auth/dto/Session.dto";
 import { OrderEntity } from "../../orders/entities/Order.entity";
 import { ApiException } from "../../common/ApiException";
 import { PaymentMethod } from "../models/PaymentMethod";
